+++ src/font.h
@@ -33,13 +33,13 @@ private:
 	std::vector<sf::Font*> fonts;
 	std::map<std::string, std::size_t> fontindex;
 
-	int    curr_font;
+	int curr_font;
 	sf::Color curr_col;
-	float  curr_size;
-	float  curr_fact;		// the length factor
+	unsigned int curr_size;
+	float curr_fact;		// the length factor
 
-	void DrawText(float x, float y, const sf::String& text, std::size_t font, float size) const;
-	void GetTextSize(const sf::String& text, float &x, float &y, std::size_t font, float size) const;
+	void DrawText(float x, float y, const sf::String& text, std::size_t font, unsigned int size) const;
+	void GetTextSize(const sf::String& text, float &x, float &y, std::size_t font, unsigned int size) const;
 public:
 	CFont();
 	~CFont();
@@ -50,29 +50,29 @@ public:
 	bool LoadFontlist();
 	std::size_t GetFontIdx(const std::string &name) const;
 	const sf::Font& getCurrentFont() const { return *fonts[curr_font]; }
-	float GetSize() const { return curr_size; }
+	unsigned int GetSize() const { return curr_size; }
 
 	// properties
-	void SetProps(const std::string &fontname, float size, const sf::Color& col);
-	void SetProps(const std::string &fontname, float size);
+	void SetProps(const std::string &fontname, unsigned int size, const sf::Color& col);
+	void SetProps(const std::string &fontname, unsigned int size);
 	void SetColor(const sf::Color& col) { curr_col = col; }
-	void SetSize(float size) { curr_size = size; }
+	void SetSize(unsigned int size) { curr_size = size; }
 	void SetFont(const std::string& fontname);
 	void SetFontFromSettings();
 
 	// auto
-	int AutoSizeN(int rel_val);	// rel_val = relative size, return: autosize
-	int AutoDistanceN(int rel_val);	// rel_val = relative dist
+	unsigned int AutoSizeN(int rel_val);	// rel_val = relative size, return: autosize
+	int AutoDistanceN(int rel_val);			// rel_val = relative dist
 
 	// draw
 	void DrawString(float x, float y, const sf::String &s) const; // sf::String class
-	void DrawString(float x, float y, const sf::String &s, const std::string &fontname, float size) const;
+	void DrawString(float x, float y, const sf::String &s, const std::string &fontname, unsigned int size) const;
 
 	// metrics
 	void  GetTextSize(const sf::String& text, float &x, float &y) const;
-	void  GetTextSize(const sf::String& text, float &x, float &y, const std::string &fontname, float size) const;
+	void  GetTextSize(const sf::String& text, float &x, float &y, const std::string &fontname, unsigned int size) const;
 	float GetTextWidth(const sf::String& text) const;
-	float GetTextWidth(const sf::String& text, const std::string &fontname, float size) const;
+	float GetTextWidth(const sf::String& text, const std::string &fontname, unsigned int size) const;
 
 	std::vector<std::string> MakeLineList(const char *source, float width);
 };
